pyproject.toml
[tool.poetry]
name = "altologicaapi"
version = "0.1.0"
description = ""
authors = ["Nahuel Rossi <nahuelrossi.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
